The most significant changes in the codebase include a refactoring of the namespace from `FleetManagement.Models` to `Shared.ApiModels`, suggesting a shift towards shared models across different projects or services. The addition of several new package references to `FleetManagement.csproj` indicates the introduction of new functionalities. The removal of `Counter` and `Weather` components and the addition of a `VehicleList` component suggest a change in the application's focus. The changes in `Program.cs` and `index.html` allow the use of Blazorise, FontAwesome icons, and Tailwind CSS throughout the application. Lastly, the `BrandModel`, `CarModel`, `VehicleModel` classes, and the `Energy` enum in the `Shared` namespace were changed from internal to public, allowing them to be used outside of their assembly.
List of changes:

1. The `FleetManagement.Models` namespace was replaced with `Shared.ApiModels` in several files, indicating a refactoring towards shared models (reference: changes in various files).
2. The vehicle list display in `VehicleListComponent.razor` was commented out and then uncommented, possibly a development error (reference: `VehicleListComponent.razor`).
3. The `GetCarsModels` and `GetBrands` methods in `VehicleData.cs` were changed from private to public and now return a list of models (reference: `VehicleData.cs`).
4. New package references were added to `FleetManagement.csproj`, providing additional functionality (reference: `FleetManagement.csproj`).
5. The `Counter` and `Weather` components were removed, and a `VehicleList` component was added, indicating a shift in application focus (reference: changes in various files).
6. Blazorise, FontAwesome icons, and Tailwind CSS providers were added to the service collection in `Program.cs` (reference: `Program.cs`).
7. A Tailwind CSS stylesheet was added to `index.html` (reference: `index.html`).
8. The `BrandModel`, `CarModel`, `VehicleModel` classes, and the `Energy` enum in the `Shared` namespace were changed from internal to public (reference: changes in various files).
