+++ FleetManagement/Components/VehicleListComponent.razor
@@ -1,8 +1,6 @@
-﻿<h3>VehicleListComponent</h3>
-
-<div class="row">
+﻿<div class="row">
     <div class="col-md-12">
-        <table class="table table-striped">
+       @*  <table class="table table-striped">
             <thead>
                 <tr>
                     <th>Id</th>
@@ -23,12 +21,12 @@
                         <td>@vehicle.Year</td>
                         <td>@vehicle.LicensePlate</td>
                         <td>
-                            @* <button class="btn btn-primary" @onclick="() => EditVehicle(vehicle.Id)">Edit</button>
-                            <button class="btn btn-danger" @onclick="() => DeleteVehicle(vehicle.Id)">Delete</button> *@
+                            <button class="btn btn-primary" @onclick="() => EditVehicle(vehicle.Id)">Edit</button>
+                            <button class="btn btn-danger" @onclick="() => DeleteVehicle(vehicle.Id)">Delete</button>
                         </td>
                     </tr>
                 }
             </tbody>
-        </table>
+        </table> *@
     </div>
     </div>
