+++ FleetManagement/Layout/NavMenu.razor
@@ -15,8 +15,8 @@
             </NavLink>
         </div>
         <div class="nav-item px-3">
-            <NavLink class="nav-link" href="counter">
-                <span class="bi bi-plus-square-fill-nav-menu" aria-hidden="true"></span> Counter
+            <NavLink class="nav-link" href="vehiclelist">
+                <span class="bi bi-plus-square-fill-nav-menu" aria-hidden="true"></span> Vehicule
             </NavLink>
         </div>
         <div class="nav-item px-3">
