--- FleetManagement/Models/CarModel.cs
@@ -1,8 +1,0 @@
-﻿namespace FleetManagement.Models
-{
-    public class CarModel
-    {
-        public int Id { get; set; }
-        public string Name { get; set; }
-    }
-}
