+++ FleetManagement/Pages/Home.razor
@@ -4,4 +4,4 @@
 
 <h1>Hello, world!</h1>
 
-<FleetManagement.Components.VehicleListComponent/>
+@* <FleetManagement.Components.VehicleListComponent/> *@
