+++ FleetManagement/_Imports.razor
@@ -8,3 +8,4 @@
 @using Microsoft.JSInterop
 @using FleetManagement
 @using FleetManagement.Layout
+@using Blazorise
