+++ FleetManagement/wwwroot/index.html
@@ -7,6 +7,7 @@
     <title>FleetManagement</title>
     <base href="/" />
     <link rel="stylesheet" href="css/bootstrap/bootstrap.min.css" />
+    <link href="_content/Blazorise.Tailwind/blazorise.tailwind.css" rel="stylesheet" />
     <link rel="stylesheet" href="css/app.css" />
     <link rel="icon" type="image/png" href="favicon.png" />
     <link href="FleetManagement.styles.css" rel="stylesheet" />
